feat: automatically migrate v2 to v3 reports
We've recently made breaking changes with
https://github.com/angular/web-codegen-scorer/commit/41ada541481a10c99de055ab6bb1c19b06a25b88,
and we want to ensure existing reports for users continue to work.

This commit lazily migrates v2 to v3 reports when loaded. We can keep
this logic for a while; while we also migrate our own Firebase database
for permanent storage (maybe we need to expose this script for users as
well eventually).